requirements.txt
faiss-cpu
numpy
pandas
requests
beautifulsoup4
tqdm
sentence-transformers
transformers
torch
openai